server: Fix sign and byte order of sent timestamp.
The timestamp sent back in RFC 868 should be encoded as a big-endian
32 bit signed integer. Use Go's stdlib's encoding/binary to convert
our timestamp into this format.

When rdate(1) was used with the previous code, it returned wrong and
wildly varying times. This patch fixes that.

Additionally, we ensure that our new use of bytes.Buffer does not
cause new memory allocations in a hot code path.